Move the Apple Sign In button to Login Carousel
As per the UI design the button is in the last page of the
login Carousel. This patch removes the unused apple-sign-in storyboard.
Also removes some unused tests. Few enums in the new stage decider
are not needed anymore
