--- ostelco-ios-client/Storyboards/SignInWithApple.storyboard
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="14854.2" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="wWg-zg-EZ0">
-    <device id="retina6_1" orientation="portrait" appearance="light"/>
-    <dependencies>
-        <deployment identifier="iOS"/>
-        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14806.4"/>
-        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
-        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
-    </dependencies>
-    <scenes>
-        <!--Sign In With Apple View Controller-->
-        <scene sceneID="0kn-5t-h2M">
-            <objects>
-                <viewController storyboardIdentifier="SignInWithAppleViewController" id="wWg-zg-EZ0" customClass="SignInWithAppleViewController" customModule="Oya_Development_app" customModuleProvider="target" sceneMemberID="viewController">
-                    <view key="view" contentMode="scaleToFill" id="33r-Au-2Ow">
-                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
-                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
-                        <subviews>
-                            <stackView opaque="NO" contentMode="scaleToFill" axis="vertical" translatesAutoresizingMaskIntoConstraints="NO" id="yZF-fe-iNH">
-                                <rect key="frame" x="20" y="747" width="374" height="70"/>
-                                <constraints>
-                                    <constraint firstAttribute="height" constant="70" id="CrZ-O9-6Rp"/>
-                                </constraints>
-                            </stackView>
-                        </subviews>
-                        <color key="backgroundColor" systemColor="systemBackgroundColor" cocoaTouchSystemColor="whiteColor"/>
-                        <constraints>
-                            <constraint firstItem="R6b-iY-t27" firstAttribute="trailing" secondItem="yZF-fe-iNH" secondAttribute="trailing" constant="20" id="0bh-Oa-OTr"/>
-                            <constraint firstItem="R6b-iY-t27" firstAttribute="bottom" secondItem="yZF-fe-iNH" secondAttribute="bottom" constant="45" id="yyU-Jx-1c5"/>
-                            <constraint firstItem="yZF-fe-iNH" firstAttribute="leading" secondItem="R6b-iY-t27" secondAttribute="leading" constant="20" id="zjD-T4-ji0"/>
-                        </constraints>
-                        <viewLayoutGuide key="safeArea" id="R6b-iY-t27"/>
-                    </view>
-                    <connections>
-                        <outlet property="buttonStackView" destination="yZF-fe-iNH" id="fwD-pu-veB"/>
-                    </connections>
-                </viewController>
-                <placeholder placeholderIdentifier="IBFirstResponder" id="nPe-RK-HUf" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
-            </objects>
-            <point key="canvasLocation" x="44.927536231884062" y="-131.25"/>
-        </scene>
-    </scenes>
-</document>
